Refractor file organization
