+++ MEArec/__init__.py
@@ -1,7 +1,8 @@
 from MEArec.tools import load_templates, load_recordings, save_recording_generator, save_template_generator, \
     get_default_config, plot_templates, plot_recordings, plot_rasters, plot_waveforms, get_templates_features, \
     plot_pca_map, get_default_recordings_params, get_default_templates_params, get_default_cell_models_folder
-from MEArec.generators import gen_recordings, gen_templates, RecordingGenerator, TemplateGenerator, SpikeTrainGenerator
+from MEArec.generation_tools import gen_recordings, gen_templates
+from MEArec.generators import RecordingGenerator, SpikeTrainGenerator, TemplateGenerator
 from MEArec.simulate_cells import return_cell, run_cell_model, return_cell_morphology
 
 from .version import version as __version__
