+++ MEArec/generators/__init__.py
@@ -1,0 +1,3 @@
+from .recordinggenerator import RecordingGenerator
+from .spiketraingenerator import SpikeTrainGenerator
+from .templategenerator import TemplateGenerator
